requirements.txt
shiny>=0.6.0
requests>=2.31.0
plotly>=5.18.0
pandas>=2.1.0
folium>=0.15.0
python-dotenv>=1.0.0
pytz>=2024.1
